SUGGESTION - AOS-1980 (#181)
* try to simplify the handler and not create objects in there.

* try to simplify the handler and not create objects in there.
