+++ src/main/kotlin/no/skatteetaten/aurora/boober/mapper/v1/AuroraDeploymentSpecBuilder.kt
@@ -1,7 +1,11 @@
 package no.skatteetaten.aurora.boober.mapper.v1
 
 import no.skatteetaten.aurora.boober.mapper.AuroraConfigFields
-import no.skatteetaten.aurora.boober.model.*
+import no.skatteetaten.aurora.boober.model.ApplicationId
+import no.skatteetaten.aurora.boober.model.AuroraConfig
+import no.skatteetaten.aurora.boober.model.AuroraConfigFile
+import no.skatteetaten.aurora.boober.model.AuroraDeploymentSpec
+import no.skatteetaten.aurora.boober.model.TemplateType
 
 @JvmOverloads
 fun createAuroraDeploymentSpec(auroraConfig: AuroraConfig, applicationId: ApplicationId,
@@ -36,7 +40,6 @@ fun createAuroraDeploymentSpec(auroraConfig: AuroraConfig, applicationId: Applic
     val handlers = applicationHandler.handlers(rawHandlers)
     val auroraConfigFields = AuroraConfigFields.create(handlers, applicationFiles)
     AuroraDeploymentSpecConfigFieldValidator(applicationId, applicationFiles, handlers, auroraConfigFields).validate()
-    applicationHandler.validate(applicationId, applicationFiles, handlers, auroraConfigFields)
 
     val volume = if (type == TemplateType.build) null else volumeMapper.createAuroraVolume(auroraConfigFields)
     val route = if (type == TemplateType.build) null else routeMapper.route(auroraConfigFields)
