+++ src/test/resources/samples/result/booberdev/sprocket/deploymentconfig.json
@@ -152,6 +152,10 @@
               {
                 "name": "MANAGEMENT_HTTP_PORT",
                 "value": "8081"
+              },
+              {
+                "name": "JOLOKIA_HTTP_PORT",
+                "value": "8778"
               }
             ],
             "imagePullPolicy": "IfNotPresent",
