+++ src/test/resources/samples/result/secretmount/aos-simple/deploymentconfig.json
@@ -157,6 +157,10 @@
               {
                 "name": "MANAGEMENT_HTTP_PORT",
                 "value": "8081"
+              },
+              {
+                "name": "JOLOKIA_HTTP_PORT",
+                "value": "8778"
               }
             ],
             "imagePullPolicy": "IfNotPresent",
